Add files via upload
Abstract - Choosing a good title for an article is an important step in the writing process. The more interesting the article title seems, the higher the chance a reader will interact with the whole content. This project focus on predicting the number of retweets and likes on Twitter from FreeCodeCamp's articles based on its titles. This problem is a classification task using Supervised Learning. With data from FreeCodeCamp on Twitter and Medium, it was used machine learning methods including support vector machines (SVM), decision trees, gaussian naive Bayes (GaussianNB), k-nearest neighbors, logistic regression, gradient boosting and naive Bayes classifier for multinomial models (MultinomialNB) to make the predictions. This study shows that the MultinomialNB model performed better for retweets reaching an accuracy of 60.6% and logistic regression reached 55.3% for likes.